myproject/requirements.txt
anthropic==0.40.0
asgiref==3.9.1
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-csp==4.0b1
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic[email]==2.10.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
reportlab==4.4.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
